### 2021-06-25 #3406 @kyleniemeyer
Hi @mikegrudic, thanks for your interest in JOSS.

Unfortunately, we have extremely limited editor availability to handle submissions right now, and so we need to put your submission on our waitlist untill an appropriate editor becomes available. Thanks for your patience!
