### 2021-07-02 0da0437
allow users to pass in tree w/o source pos and masses, force method = tree in this case
