### 2021-07-27 #3406 @mikegrudic
ngoldbaum and

> @whedon assign me as editor
>
> Sure - I'm happy to edit! I'm a bit behind on things so I'll take a first pass of the paper and repo around the middle of next week, and then start recruiting reviewers.
>
> In the meantime, @mikegrudic, let me know if you have any reviewer recommendations from our [main list of volunteers](https://bit.ly/joss-reviewers) or if there's anyone else who you think would be particularly well suited (include their github handle, but don't @ them because then they'll get inundated with notifications :D).

Thanks! On the list, some potential reviewers I see who would likely be knowledgeable about the algorithms and applications of the package would be ngoldbaum and zingale.
