### 2021-07-28 8e381c9
added support for computing correlation functions
