+++ src/pytreegrav/dynamic_tree.py
@@ -1,0 +1,213 @@
+from numba import int32, deferred_type, optional, float64, boolean, int64, njit, jit, prange, types
+from numba.experimental import jitclass
+import numpy as np
+from numpy import empty, empty_like, zeros, zeros_like, sqrt, ones
+
+spec = [
+    ('Sizes', float64[:]), # side length of tree nodes
+    ('Deltas', float64[:]), # distance between COM and geometric center of node
+    ('Coordinates', float64[:,:]), # location of center of mass of node (actually stores _geometric_ center before we do the moments pass)
+    ('Velocities', float64[:,:]), # velocity of the center of mass of node
+    ('VelocityDisp',float64[:]), # center-of-mass velocity dispersion 
+    ('Masses', float64[:]), # total mass of node
+    ('Quadrupoles', float64[:,:,:]), # Quadrupole moment of the node
+    ('HasQuads', boolean), # Allow us to quickly check if Quadrupole moments exist to keep monopole calculations fast
+    ('NumParticles', int64), # number of particles in the tree
+    ('NumNodes',int64), # number of particles + nodes (i.e. mass elements) in the tree
+    ('Softenings', float64[:]), # individual softenings for particles, _maximum_ softening of inhabitant particles for nodes
+    ('NextBranch',int64[:]), 
+    ('FirstSubnode',int64[:]),
+    ('TreewalkIndices',int64[:]),
+   # ('children',int64[:,:]) # indices of child nodes
+]
+
+
+octant_offsets = 0.25 * np.array([[-1,-1,-1],
+                                  [1,-1,-1],
+                                  [-1,1,-1],
+                                  [1,1,-1],
+                                  [-1,-1,1],
+                                  [1,-1,1],
+                                  [-1,1,1],
+                                  [1,1,1]])
+
+@jitclass(spec)
+class DynamicOctree(object):
+    """Octree implementation that stores node velocities for correlation functions and dynamic updates."""
+
+    def __init__(self, points, masses, softening, vels, morton_order=True, quadrupole=False):
+        self.TreewalkIndices = -ones(points.shape[0],dtype=np.int64)        
+        self.HasQuads = quadrupole
+        children = self.BuildTree(points, masses, softening, vels) # first provisional treebuild to get the ordering right
+        SetupTreewalk(self,self.NumParticles,children) # set up the order of the treewalk
+        ComputeMomentsDynamic(self,self.NumParticles,children) # compute centers of mass, etc.         
+        self.GetWalkIndices() # get the Morton ordering of the points
+ 
+        if morton_order: # if enabled, we rebuild the tree in Morton order (the order that points are visited in the depth-first traversal)
+            children = self.BuildTree(points[self.TreewalkIndices], np.take(masses,self.TreewalkIndices), np.take(softening,self.TreewalkIndices), vels[self.TreewalkIndices]) # now re-build the tree with everything in order
+            SetupTreewalk(self,self.NumParticles,children) # re-do the treewalk order with the new indices
+
+        ComputeMomentsDynamic(self,self.NumParticles,children) # compute centers of mass, etc.
+            
+    def BuildTree(self,points,masses,softening,vels):
+        # initialize all attributes
+        self.NumParticles = points.shape[0]
+        self.NumNodes = 2*self.NumParticles # this is the number of elements in the tree, whether nodes or particles. can make this smaller but this has a safety factor
+        self.Sizes = zeros(self.NumNodes)
+        self.Deltas = zeros(self.NumNodes)
+        self.Masses = zeros(self.NumNodes)
+        if self.HasQuads:
[…]
