### 2021-07-29 c2cebe4
added option for weighted binning scheme when tree nodes span multiple bins in correlation functions - can reduce certain binning artifacts due to the tree structure
