+++ src/pytreegrav/treewalk.py
@@ -1,4 +1,4 @@
-from numpy import sqrt, empty, zeros, empty_like, zeros_like, dot
+from numpy import sqrt, empty, zeros, empty_like, zeros_like, dot, fabs
 from numba import njit, prange, get_num_threads
 from math import copysign
 from .kernel import *
@@ -14,11 +14,9 @@ def NearestImage(x,boxsize):
 @njit(fastmath=True)
 def PotentialWalk(pos,  tree, softening=0, no=-1, theta=0.7):
     """Returns the gravitational potential at position x by performing the Barnes-Hut treewalk using the provided octree instance
-
     Arguments:
     pos - (3,) array containing position of interest
     tree - octree object storing the tree structure    
-
     Keyword arguments:
     softening - softening radius of the particle at which the force is being evaluated - we use the greater of the target and source softenings when evaluating the softened potential
     no - index of the top-level node whose field is being summed - defaults to the global top-level node, can use a subnode in principle for e.g. parallelization
@@ -55,11 +53,9 @@ def PotentialWalk(pos,  tree, softening=0, no=-1, theta=0.7):
 @njit(fastmath=True)
 def PotentialWalk_quad(pos,  tree, softening=0, no=-1, theta=0.7):
     """Returns the gravitational potential at position x by performing the Barnes-Hut treewalk using the provided octree instance. Uses the quadrupole expansion.
-
     Arguments:
     pos - (3,) array containing position of interest
     tree - octree object storing the tree structure    
-
     Keyword arguments:
     softening - softening radius of the particle at which the force is being evaluated - we use the greater of the target and source softenings when evaluating the softened potential
     no - index of the top-level node whose field is being summed - defaults to the global top-level node, can use a subnode in principle for e.g. parallelization
@@ -102,11 +98,9 @@ def PotentialWalk_quad(pos,  tree, softening=0, no=-1, theta=0.7):
 @njit(fastmath=True)
 def AccelWalk(pos,  tree, softening=0, no=-1, theta=0.7): #,include_self_potential=False):
     """Returns the gravitational acceleration field at position x by performing the Barnes-Hut treewalk using the provided octree instance
-
     Arguments:
     pos - (3,) array containing position of interest
     tree - octree instance storing the tree structure    
-
     Keyword arguments:
     softening - softening radius of the particle at which the force is being evaluated - we use the greater of the target and source softenings when evaluating the softened potential
     no - index of the top-level node whose field is being summed - defaults to the global top-level node, can use a subnode in principle for e.g. parallelization
@@ -150,11 +144,9 @@ def AccelWalk(pos,  tree, softening=0, no=-1, theta=0.7): #,include_self_potenti
 @njit(fastmath=True)
 def AccelWalk_quad(pos,  tree, softening=0, no=-1, theta=0.7): #,include_self_potential=False):
     """Returns the gravitational acceleration field at position x by performing the Barnes-Hut treewalk using the provided octree instance. Uses the quadrupole expansion.
-
     Arguments:
     pos - (3,) array containing position of interest
[…]
