### 2021-07-29 f91b0c6
ComputeCorrFuncs.py uses weighted binning scheme
