+++ scripts/ComputeCorrFuncs.py
@@ -24,12 +24,12 @@ for f in argv[1:]:
         boxsize = F["Header"].attrs["BoxSize"]
 
         rbins = np.logspace(-3,2,61)
-        rbins, mbin = DensityCorrFunc(x,m,rbins,theta=0.5,parallel=True,boxsize=boxsize)
+        rbins, mbin = DensityCorrFunc(x,m,rbins,theta=0.7,parallel=True,boxsize=boxsize,weighted_binning=True)
         rhocorr = mbin / np.diff(4*np.pi*rbins**3 / 3) # convert from mass in bins to average density profile around a point
-        rbins, Sv = VelocityStructFunc(x,vol,v,rbins,theta=0.5,parallel=True,boxsize=boxsize)
-        rbins, SB = VelocityStructFunc(x,vol,B,rbins,theta=0.5,parallel=True,boxsize=boxsize)
-        rbins, vcorr = VelocityCorrFunc(x,vol,v,rbins,theta=0.5,parallel=True,boxsize=boxsize)
-        rbins, Bcorr = VelocityCorrFunc(x,vol,B,rbins,theta=0.5,parallel=True,boxsize=boxsize)
+        rbins, Sv = VelocityStructFunc(x,vol,v,rbins,theta=0.7,parallel=True,boxsize=boxsize,weighted_binning=True)
+        rbins, SB = VelocityStructFunc(x,vol,B,rbins,theta=0.7,parallel=True,boxsize=boxsize,weighted_binning=True)
+        rbins, vcorr = VelocityCorrFunc(x,vol,v,rbins,theta=0.7,parallel=True,boxsize=boxsize,weighted_binning=True)
+        rbins, Bcorr = VelocityCorrFunc(x,vol,B,rbins,theta=0.7,parallel=True,boxsize=boxsize,weighted_binning=True)
 
         r_avg = (rbins[1:]*rbins[:-1])**0.5
         np.savetxt(corrpath + "/corrfunc_" + snapnum + ".dat", np.c_[r_avg, rhocorr, vcorr, Bcorr, Sv, SB], header="#COLUMNS: (0) bin effective radius (1) density autocorrelation (2) velocity autocorrelation (3) Magnetic field autocorrelation (4) Velocity structure function (5) Magnetic field structure function" )
