### 2021-07-29 b6cc68b
randomize binning of nodes to mitigate artifacts
