### 2021-08-31 #3675 @whedon
[REVIEW]: pytreegrav: a fast Python gravity solver
**Submitting author:** @mikegrudic (<a href="http://orcid.org/0000-0002-1655-5604">Michael Grudi´c</a>)
**Repository:** <a href="https://github.com/mikegrudic/pytreegrav" target ="_blank">https://github.com/mikegrudic/pytreegrav</a>
**Version:** 1.0.0

- [x] #3785
