### 2021-09-25 #3675 @adrn
@mikegrudic The submitted paper is excellent: concise, clearly written, and I don't have any comments or edits to suggest - nice work! Comments on the code are in: mikegrudic/pytreegrav#12
