### 2021-09-28 #3675 @maxwelltsai
@dfm @whedon It seems that the paper link is broken. I can no longer download the paper proof from this page: https://github.com/openjournals/joss-reviews/issues/3406#issuecomment-868250511

Has it been transferred to somewhere else? I did find the `paper.md` file on the code repo, but I am wondering whether this is the paper itself.
