### 2021-09-28 #3675 @dfm
@maxwelltsai: There is a new link just a few comments up ☝️: https://github.com/openjournals/joss-reviews/issues/3675#issuecomment-909502893 That one's working for me - does it work for you too?
