### 2021-11-22 aa4e865
remove scripts/ComputeCorrFuncs.py (now available in starforge_tools)
