+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. pytreegrav documentation master file, created by
+   sphinx-quickstart on Mon Nov 22 10:52:56 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to pytreegrav's documentation!
+======================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
