### 2021-11-22 395c1df
setup auto-installs dependencies
