+++ setup.py
@@ -1,8 +1,15 @@
-import setuptools
+import setuptools, os
 
 with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
+thelibFolder = os.path.dirname(os.path.realpath(__file__))
+requirementPath = thelibFolder + '/requirements.txt'
+install_requires = []
+if os.path.isfile(requirementPath):
+    with open(requirementPath) as f:
+        install_requires = f.read().splitlines()
+    
 setuptools.setup(
     name="pytreegrav",
     version="0.27",
@@ -22,4 +29,5 @@ setuptools.setup(
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
     python_requires=">=3.6",
+    install_requires=install_requires
 )
