+++ docs/source/index.rst
@@ -10,6 +10,8 @@ Welcome to pytreegrav's documentation!
    :maxdepth: 2
    :caption: Contents:
 
+   usage/installation
+   usage/quickstart
 
 
 Indices and tables
