+++ src/pytreegrav/frontend.py
@@ -19,14 +19,21 @@ def valueTestMethod(method):
         raise ValueError("Invalid method %s. Must be one of: %s"%(method,str(methods)))
 
 def ConstructTree(pos,m,softening=None,quadrupole=False,vel=None):
-    """Builds the tree containing particle data, for subsequent potential/field evaluation
-    Arguments:
-    pos -- shape (N,3) array of particle positions
-    m -- shape (N,) array of particle masses
-    softening -- shape (N,) array of particle softening lengths
+    """Builds a tree containing particle data, for subsequent potential/field evaluation
 
-    Returns:
-    Octree instance built from particle data
+    Parameters
+    ----------
+    pos: array_like
+        shape (N,3) array of particle positions
+    m: array_like
+        shape (N,) array of particle masses
+    softening: array_like or None, optional
+        shape (N,) array of particle softening lengths
+
+    Returns
+    -------
+    tree: octree
+        Octree instance built from particle data
     """
     if softening is None: softening = zeros_like(m)
     if not (np.all(np.isfinite(pos)) and np.all(np.isfinite(m)) and np.all(np.isfinite(softening))):
@@ -38,24 +45,39 @@ def ConstructTree(pos,m,softening=None,quadrupole=False,vel=None):
         return DynamicOctree(pos, m,softening,vel, quadrupole=quadrupole)
 
 def Potential(pos, m, softening=None, G=1., theta=.7, tree=None, return_tree=False,parallel=False,method='adaptive',quadrupole=False):
-    """Returns the gravitational potential for a set of particles with positions x and masses m, at the positions of those particles, using either brute force or tree-based methods depending on the number of particles.
+    """Gravitational potential calculation
 
-    Arguments:
-    pos -- shape (N,3) array of particle positions
-    m -- shape (N,) array of particle masses
+    Returns the gravitational potential for a set of particles with positions x and masses m, at the positions of those particles, using either brute force or tree-based methods depending on the number of particles.
 
-    Optional arguments:
-    G -- gravitational constant (default 1.0)
-    softening -- shape (N,) array containing kernel support radii for gravitational softening - set to 0 by default
-    theta -- cell opening angle used to control force accuracy; smaller is slower (runtime ~ theta^-3) but more accurate. (default 0.7, gives ~1% accuracy)
-    parallel -- If True, will parallelize the force summation over all available cores. (default False)
-    tree -- optional pre-generated Octree: this can contain any set of particles, not necessarily the target particles at pos (default None)
-    return_tree -- return the tree used for future use (default False)
-    method -- 'adaptive', 'tree', or 'bruteforce' (default adaptive tries to pick the faster choice)
+    Parameters
+    ----------
+    pos: array_like
+        shape (N,3) array of particle positions
+    m: array_like
+        shape (N,) array of particle masses
+    G: float, optional
+        gravitational constant (default 1.0)
+    softening: None or array_like, optional
+        shape (N,) array containing kernel support radii for gravitational softening - set to 0 by default
+    theta: float, optional
+        cell opening angle used to control force accuracy; smaller is slower (runtime ~ theta^-3) but more accurate. (default 0.7, gives ~1% accuracy)
+    parallel: bool, optional
+        If True, will parallelize the force summation over all available cores. (default False)
+    tree: Octree, optional
+        optional pre-generated Octree: this can contain any set of particles, not necessarily the target particles at pos (default None)
+    return_tree: bool, optional
+        return the tree used for future use (default False)
+    method: str, optional
+        Which summation method to use: 'adaptive', 'tree', or 'bruteforce' (default adaptive tries to pick the faster choice)
+    quadrupole: bool, optional
[…]
