+++ docs/source/frontend_API.rst
@@ -1,0 +1,6 @@
+API Documentation
+=================
+
+.. automodule:: pytreegrav.frontend
+   :noindex:
+   :members:
