+++ docs/source/pytreegrav.rst
@@ -1,0 +1,61 @@
+pytreegrav package
+==================
+
+Submodules
+----------
+
+pytreegrav.bruteforce module
+----------------------------
+
+.. automodule:: pytreegrav.bruteforce
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pytreegrav.dynamic\_tree module
+-------------------------------
+
+.. automodule:: pytreegrav.dynamic_tree
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pytreegrav.frontend module
+--------------------------
+
+.. automodule:: pytreegrav.frontend
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pytreegrav.kernel module
+------------------------
+
+.. automodule:: pytreegrav.kernel
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pytreegrav.octree module
+------------------------
+
+.. automodule:: pytreegrav.octree
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pytreegrav.treewalk module
+--------------------------
+
+.. automodule:: pytreegrav.treewalk
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pytreegrav
+   :members:
+   :undoc-members:
+   :show-inheritance:
