+++ requirements.txt
@@ -1,0 +1,2 @@
+numba
+numpy<1.21
