+++ requirements.txt
@@ -1,2 +1,3 @@
 numba
 numpy<1.21
+pytreegrav
