+++ requirements.txt
@@ -1,3 +1,2 @@
 numba
 numpy<1.21
-pytreegrav
