+++ docs/build/html/_sources/modules.rst.txt
@@ -1,0 +1,7 @@
+src
+===
+
+.. toctree::
+   :maxdepth: 4
+
+   pytreegrav
