+++ docs/build/html/_sources/index.rst.txt
@@ -13,6 +13,7 @@ pytreegrav is a package for computing the gravitational potential and/or field o
 
    usage/installation
    usage/quickstart
+   Nbody_simulation
    frontend_API
 
 Indices and tables
