+++ requirements.txt
@@ -1,2 +1,2 @@
-numba
+numba<0.54
 numpy<1.21
