### 2021-11-23 #3675 @mikegrudic
Thanks to @maxwelltsai and @adrn for taking the time to give useful feedback. The package has been revised extensively, and in particular we now have [sphinx documentation](https://pytreegrav.readthedocs.io/en/latest/) and an automated `pytest` test. This contains a lot of the additional content requested in your reviews. There is still a ways to go to fully document the codebase, but at least all of the frontend API functions are fully documented now. Further feedback on this is welcome.

Responding to some specific points apart from the general suggestion of expanded documentation, first by @adrn:

>     * I don't see any mention of what profile is used to soften the point mass potentials when `h` is provided. Is there a standard profile to use (i.e. Plummer?), or is this configurable? It would be good to explicitly state this in the documentation.
>

We mention in the paper that the softened force is that of a cubic M4 spline mass distribution and provide the reference for the specific mathematical expressions. We now also clarify this in the API documentation, where the respective softening arguments are documented.

>     * Though the details of code style can be subjective, Python packages generally at least follow the convention (laid out in PEP 8) that class names follow [the CapWords convention](https://www.python.org/dev/peps/pep-0008/#class-names) and [function names are lower case](https://www.python.org/dev/peps/pep-0008/#id43). Please consider reformatting the functions in this package to follow PEP8 guidelines.

In a vacuum I'd like to do this, but at this point I am disinclined for fear of causing headaches and confusion for the various projects that are already using this package, many of which I do not own.

 And @maxwelltsai's points:

> @mikegrudic, @dfm I have finished the review. Apart from the comments from @adrn, I have the following comments:
>
>     * I would suggest that the authors add the dependent packages (e.g., `numpy`, `numba`) into their `setup.py`. This allows the dependency to automatically sorted out when the users install the package.
>
Done.

>     * I assume that the library works with any version of Python 3.x. If the package requires a specific version of Python, it would be good to document this as well.
>     * I notice that `pytreegrav` is already on PyPI. I would also suggest the authors to document in their GitHub repository (and also mention in the paper) that the package can be installed using PyPI: `pip install pytreegrav`. Some users may find it easier to directly install the package with `pip` rather than having to clone the code and run `setup.py`.
>

Done, see the installation section of the documentation.

>     * The package is designed for the specific purpose of calculating the potential and accelerations. I understand that it is a software library to be integrated into a Python N-body integrator, but it would be great if the authors could demonstrate this by implementing a simple integrator (e.g., LeapFrog) with `pytreegrav`. This can be done within a Jupyter notebook and/or a standalone python script. I would also suggest that the authors to add a directory `examples/` in the root of their repository, which contains all these examples for new users to get started.
>
Done, see `examples/Nbody_integrator.ipynb`

>     * Since the authors use the JIT decorator in `numpy` to parallelise the computation on CPUs, is there any reason that they didn't try to use `cuda.jit` for GPU parallelisation?

Just that I've never personally been able to get numba's cuda interface working on any system of my own 🤷‍♂️ this is probably possible, because the tree structure is really a set of float/int arrays for which the Octree instance serves as a wrapper, so one could write a kernel that just accepts the unpacked arrays as arguments, and otherwise run the same loop as in the CPU treewalk functions. I encourage anyone interested to give this a shot.
