+++ docs/source/community.rst
@@ -1,0 +1,4 @@
+Feedback, Support, and Contributions
+====================================
+
+To contribute to pytreegrav, report an issue, or seek support, please initiate a pull request or issue through the project `project github <https://github.com/mikegrudic/pytreegrav>`_
