### 2021-12-17 #3675 @dfm
@adrn, @maxwelltsai: Thanks for your updates. Can you both go through [your checklists up at the top of this thread](https://github.com/openjournals/joss-reviews/issues/3675#issue-984176521), because I'm still seeing a lot of un-checked boxes. Thanks!
