### 2021-12-27 #3675 @dfm
@mikegrudic I've just opened [a small PR](https://github.com/mikegrudic/pytreegrav/pull/13) over on your repo. Once you've checked that out and merged, here are a few last steps for you, then we'll be ready to go!

1. Comment `@whedon generate pdf`, then take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations. I've taken a pass and it looks good to me!
3. Create an archived release of that version of the software (using Zenodo or something similar). **Please make sure that the metadata (title and author list) exactly match the paper.** Then report the DOI of the release back to this thread.
