### 2021-12-28 #3675 @mikegrudic
@dfm thanks for your work on this. I have incremented the version to 1.0.0, and have uploaded the version to zenodo: 10.5281/zenodo.5807909
