### 2021-12-30 #3675 @arfon
Also, I meant to say that this was JOSS' 1500th published paper! 🎉
